Consolidate to a single CLI: cmd/stochadex-full -> cmd/stochadex
The engine module carried a lean cmd/stochadex (engine + Postgres) that was
never the shipped artifact — releases, the Docker image, and the plugin all use
the separate cmd/stochadex-full module. Its only unique value was a niche lean
go-install / WASM CLI, none of it in the shipping path. CGO_ENABLED=0 does not
reduce -full to the lean CLI either: Arrow and S3 are pure Go and always
compiled in, so the two genuinely differed by Arrow/S3, not just cgo.

Collapse to one CLI:
- Rename the cmd/stochadex-full module to cmd/stochadex (module path
  github.com/umbralcalc/stochadex/cmd/stochadex) and delete the former lean
  engine-module CLI. The engine module is now library-only.
- The single CLI stays a nested module so its heavy, opt-in deps (Arrow, S3,
  DuckDB, ONNX Runtime) stay out of the engine's go.mod; tiered by build flags —
  pure Go (engine + Postgres + Arrow + S3) or -tags (cblas / duckdb_arrow / onnx).
- Update CI, release workflow, Dockerfile, .gitignore, and the package doc
  comments that referenced cmd/stochadex-full. Engine-library WASM-cleanliness
  (what dexetera-style in-browser embedding relies on) is unaffected.
- CHANGELOG: document this plus the ONNX inference partition under Unreleased;
  the module-path rename is breaking, noted as such.
